feat: add lobby countdown management by integrating countdown state into app state and updating game state handling

+++ src/models/lobby.rs
@@ -89,7 +89,7 @@ pub enum LobbyServerMessage {
     },
     NotifyKicked,
     Countdown {
-        time: u64,
+        time: u32,
     },
     GameState {
         state: GameState,
